edit category

+++ PMAPI/PMAPI/apps/transaction/models.py
@@ -107,30 +107,57 @@ class Transaction(Metadata):
         self.installment = installment
         self.save()
 
-    def update(self, single=True, future=False, all=False, *args, **kwargs):
-        print(kwargs)
+    def update(self, single=False, pending=False, all=False, *args, **kwargs):
+        to_update = []
 
-    def destroy(self, single=True, future=False, all=False, *args, **kwargs):
+        if single:
+            to_update = [self]
+
+        if pending:
+            if self.installment:
+                to_update = self.installment.get_pending_transactions()
+            if self.recurrent:
+                to_update = self.recurrent.get_pending_transactions()
+                # self.recurrent.delete()
+
+        if all:
+            if self.installment:
+                to_update = self.installment.get_all_transactions()
+            if self.recurrent:
+                to_update = self.recurrent.get_all_transactions()
+                # self.recurrent.delete()
+
+        for item in to_update:
+            # update
+            pass
+
+        return True
+
+    def destroy(self, single=False, pending=False, all=False, *args, **kwargs):
         to_delete = []
 
         if single:
             to_delete = [self]
 
-        if future:
+        if pending:
             if self.installment:
-                to_delete = self.installment.get_future_transactions(self)
+                to_delete = list(self.installment.get_pending_transactions())
             if self.recurrent:
-                to_delete = self.recurrent.get_future_transactions(self)
-                self.recurrent.delete()
+                to_delete = list(self.recurrent.get_pending_transactions())
+                to_delete.append(self.recurrent)
 
         if all:
             if self.installment:
-                to_delete = self.installment.get_all_transactions()
+                to_delete = list(self.installment.get_all_transactions())
             if self.recurrent:
-                to_delete = self.recurrent.get_all_transactions()
-                self.recurrent.delete()
+                to_delete = list(self.recurrent.get_all_transactions())
+                to_delete.append(self.recurrent)
+
+        if len(to_delete) == 0:
+            return False
 
         for item in to_delete:
+            print('deleting', item)
             item.delete()
 
         return True
